improvement: split service student controller
