+++ apps/api/student/src/app/app.controller.ts
@@ -7,31 +7,31 @@ import {
   IDeleteStudentRequest,
   IUpdateStudentRequest,
   TGraduationStatusRequest,
-  IGetPaymentObligationsRequest,
+  TPaymentObligationsRequest,
 } from "@uninus/entities";
 @Controller()
 export class AppController {
   constructor(private readonly studentService: AppService) {}
   @MessagePattern("get_student")
-  async getStudent(payload: IGetStudentRequest) {
-    return await this.studentService.getStudent(payload);
+  async getDataStudent(payload: IGetStudentRequest) {
+    return await this.studentService.getDataStudent(payload);
   }
 
   @MessagePattern("update_student")
-  async updateStudent(payload: IUpdateStudentRequest) {
-    return await this.studentService.updateStudent(payload);
+  async updateDataStudent(payload: IUpdateStudentRequest) {
+    return await this.studentService.updateDataStudent(payload);
   }
 
   @MessagePattern("delete_student")
-  async deleteStudent(payload: IDeleteStudentRequest) {
-    return await this.studentService.deleteStudent(payload);
+  async deleteDataStudent(payload: IDeleteStudentRequest) {
+    return await this.studentService.deleteDataStudent(payload);
   }
   @MessagePattern("get_graduation_status")
   async getGraduationStatus(payload: TGraduationStatusRequest) {
     return await this.studentService.checkGraduationStatus(payload);
   }
   @MessagePattern("get_payment_obligations")
-  async getPaymentObligations(payload: IGetPaymentObligationsRequest) {
+  async getPaymentObligations(payload: TPaymentObligationsRequest) {
     return await this.studentService.getPaymentObligations(payload);
   }
 }
