+++ libs/api/services/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./auth";
+export * from "./student";
 export * from "./prisma";
