+++ libs/entities/src/types/auth/index.ts
@@ -1,7 +1,7 @@
-export * from "./password.types";
-export * from "./login.types";
-export * from "./register.types";
-export * from "./logout.types";
-export * from "./otp.types";
-export * from "./register.types";
-export * from "./token.types";
+export * from "./password.type";
+export * from "./login.type";
+export * from "./register.type";
+export * from "./logout.type";
+export * from "./otp.type";
+export * from "./register.type";
+export * from "./token.type";
