+++ libs/entities/src/types/auth/logout.type.ts
