+++ libs/entities/src/types/auth/otp.type.ts
