+++ libs/entities/src/types/auth/password.type.ts
