+++ libs/entities/src/types/auth/register.type.ts
