+++ libs/entities/src/types/auth/token.type.ts
