+++ libs/entities/src/types/student/graduation.type.ts
@@ -1,0 +1,11 @@
+export type TGraduationStatusRequest = {
+  registration_number: string;
+};
+export type TGraduationStatusReponse = {
+  registration_number?: string | null;
+  fullname?: string | null;
+  department?: string | null;
+  selection_path?: string | null;
+  registration_status?: string | null;
+  message?: string;
+};
