+++ libs/entities/src/types/student/payment.type.ts
@@ -1,0 +1,10 @@
+export type TPaymentObligationsRequest = {
+  search?: string;
+  userId?: string;
+};
+export type TPaymentObligationsResponse = Array<{
+  id?: number;
+  name?: string;
+  amount?: number;
+  spelled_out?: string;
+}>;
